+++ Changelog.md
@@ -3,6 +3,7 @@
 ### 2.5 WIP
 - Pause videos and audio when changing page or tab (disabled by default for audio)
 - New option: Pause playing audio if it goes invisible
+- New option: Behave in browser like inside PWA (no elements selection, custom context menu everywhere, etc)
 
 ### 2.4 - Minimax H3 support update
 - Fixed subgraphs in new format used in Minimax H3 I2V default workflow